Restore Minizip in `ReadiumShared`
