+++ Cartfile
@@ -1,3 +1,4 @@
+github "dexman/Minizip" ~> 1.4.0
 github "krzyzanowskim/CryptoSwift" ~> 1.8.0
 github "ra1028/DifferenceKit" ~> 1.3.0
 github "readium/Fuzi" ~> 4.0.0
