+++ Sources/Shared/Toolkit/ZIP/ZIPArchiveOpener.swift
@@ -7,16 +7,11 @@
 import Foundation
 
 /// An ``ArchiveOpener`` for ZIP resources.
-public class ZIPArchiveOpener: ArchiveOpener {
-    private let opener = ZIPFoundationArchiveOpener()
-
-    public init() {}
-
-    public func open(resource: any Resource, format: Format) async -> Result<ContainerAsset, ArchiveOpenError> {
-        await opener.open(resource: resource, format: format)
-    }
-
-    public func sniffOpen(resource: any Resource) async -> Result<ContainerAsset, ArchiveSniffOpenError> {
-        await opener.sniffOpen(resource: resource)
+public class ZIPArchiveOpener: CompositeArchiveOpener {
+    public init() {
+        super.init([
+            MinizipArchiveOpener(),
+            ZIPFoundationArchiveOpener(),
+        ])
     }
 }
