+++ Support/Carthage/.xcodegen
@@ -439,9 +439,13 @@
 ../../Sources/Navigator/EPUB/Assets/Static/scripts
 ../../Sources/Navigator/EPUB/Assets/Static/scripts/.gitignore
 ../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed-wrapper-one.js
+../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed-wrapper-one.js.map
 ../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed-wrapper-two.js
+../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed-wrapper-two.js.map
 ../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed.js
+../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-fixed.js.map
 ../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-reflowable.js
+../../Sources/Navigator/EPUB/Assets/Static/scripts/readium-reflowable.js.map
 ../../Sources/Navigator/EPUB/CSS
 ../../Sources/Navigator/EPUB/CSS/CSSLayout.swift
 ../../Sources/Navigator/EPUB/CSS/CSSProperties.swift
@@ -785,8 +789,14 @@
 ../../Sources/Shared/Toolkit/XML/Fuzi.swift
 ../../Sources/Shared/Toolkit/XML/XML.swift
 ../../Sources/Shared/Toolkit/ZIP
+../../Sources/Shared/Toolkit/ZIP/Minizip
+../../Sources/Shared/Toolkit/ZIP/Minizip/MinizipArchiveOpener.swift
+../../Sources/Shared/Toolkit/ZIP/Minizip/MinizipContainer.swift
 ../../Sources/Shared/Toolkit/ZIP/ZIPArchiveOpener.swift
-../../Sources/Shared/Toolkit/ZIP/ZIPFoundation.swift
+../../Sources/Shared/Toolkit/ZIP/ZIPFoundation
+../../Sources/Shared/Toolkit/ZIP/ZIPFoundation/ZIPFoundationArchiveFactory.swift
+../../Sources/Shared/Toolkit/ZIP/ZIPFoundation/ZIPFoundationArchiveOpener.swift
+../../Sources/Shared/Toolkit/ZIP/ZIPFoundation/ZIPFoundationContainer.swift
 ../../Sources/Streamer
 ../../Sources/Streamer/Assets
 ../../Sources/Streamer/Assets/fonts
